+++ src/main/java/ma/vi/esql/builder/CreateTableBuilder.java
@@ -4,21 +4,16 @@
 
 package ma.vi.esql.builder;
 
-import ma.vi.base.lang.Builder;
-import ma.vi.esql.semantic.type.Types;
 import ma.vi.esql.syntax.Context;
 import ma.vi.esql.syntax.Parser;
-import ma.vi.esql.syntax.define.Attribute;
 import ma.vi.esql.syntax.define.Metadata;
 import ma.vi.esql.syntax.define.table.*;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static java.util.Collections.singletonList;
-import static java.util.stream.Collectors.toList;
 
 /**
  * A builder for directly constructing `create table` ESQL statements. This
@@ -27,9 +22,9 @@ import static java.util.stream.Collectors.toList;
  *
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
-public class CreateTableBuilder implements Builder<CreateTable> {
+public class CreateTableBuilder extends CreateBuilder<CreateTable, CreateTableBuilder> {
   public CreateTableBuilder(Context context) {
-    this.context = context;
+    super(context);
   }
 
   @Override
@@ -37,53 +32,8 @@ public class CreateTableBuilder implements Builder<CreateTable> {
     return new CreateTable(context, name, dropUndefined, columns, constraints, new Metadata(context, metadata));
   }
 
-  public CreateTableBuilder name(String name) {
-    this.name = name;
-    return this;
-  }
-
   public CreateTableBuilder dropUndefined(boolean dropUndefined) {
     this.dropUndefined = dropUndefined;
-    return this;
-  }
-
-  public CreateTableBuilder metadata(String name, String expression) {
-    Parser parser = new Parser(context.structure);
-    this.metadata.add(new Attribute(context, name, parser.parseExpression(expression)));
-    return this;
-  }
-
-  public CreateTableBuilder column(String name, String type, Attr... metadata) {
-    return column(name, type, false, metadata);
-  }
-
-  public CreateTableBuilder column(String name, String type, boolean notNull, Attr... metadata) {
-    return column(name, type, notNull, null, metadata);
-  }
-
-  public CreateTableBuilder column(String name, String type, boolean notNull,
-                                   String defaultExpression, Attr... metadata) {
-    Parser parser = new Parser(context.structure);
-    this.columns.add(
-        new ColumnDefinition(context, name, Types.typeOf(type),
-                             notNull, defaultExpression == null
-                     ? null
-                     : parser.parseExpression(defaultExpression),
-                             new Metadata(context,
-                Stream.of(metadata)
-                      .map(a -> new Attribute(context, a.name(), parser.parseExpression(a.expr())))
-                      .collect(toList()))));
-    return this;
-  }
-
-  public CreateTableBuilder derivedColumn(String name, String expression, Attr... metadata) {
-    Parser parser = new Parser(context.structure);
-    this.columns.add(
-        new DerivedColumnDefinition(context, name, parser.parseExpression(expression),
-                                    new Metadata(context,
-                Stream.of(metadata)
-                      .map(a -> new Attribute(context, a.name(), parser.parseExpression(a.expr())))
-                      .collect(toList()))));
     return this;
   }
 
@@ -181,15 +131,7 @@ public class CreateTableBuilder implements Builder<CreateTable> {
     return this;
   }
 
-  private String name;
-
   private boolean dropUndefined = false;
 
-  private final List<ColumnDefinition> columns = new ArrayList<>();
-
   private final List<ConstraintDefinition> constraints = new ArrayList<>();
-
-  private final List<Attribute> metadata = new ArrayList<>();
-
-  public final Context context;
 }