+++ src/main/java/ma/vi/esql/database/init/TableInitializer.java
@@ -1,14 +1,11 @@
 package ma.vi.esql.database.init;
 
-import ma.vi.esql.builder.Attr;
-import ma.vi.esql.builder.Attributes;
 import ma.vi.esql.builder.CreateTableBuilder;
 import ma.vi.esql.database.Database;
-import ma.vi.esql.database.EsqlConnection;
 import ma.vi.esql.semantic.type.BaseRelation;
 import ma.vi.esql.syntax.Context;
+import ma.vi.esql.syntax.define.table.CreateTable;
 
-import java.util.LinkedHashMap;
 import java.util.Map;
 
 import static ma.vi.esql.builder.Attributes.*;
@@ -44,72 +41,27 @@ import static ma.vi.esql.builder.Attributes.*;
  *
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
-public class TableInitializer implements Initializer<BaseRelation> {
+public class TableInitializer extends CreateInitializer<BaseRelation,
+                                                        CreateTable,
+                                                        CreateTableBuilder> {
   @Override
-  public BaseRelation add(Database     db,
-                          boolean      overwrite,
-                          String       name,
-                          BaseRelation existing,
-                          Map<String, Object> definition) {
-    CreateTableBuilder builder = new CreateTableBuilder(new Context(db.structure()));
-    builder.name(name);
-    for (var e: definition.entrySet()) {
-      String columnName = e.getKey();
-      if (columnName.equals(METADATA)) {
-        /*
-         * Table-level metadata.
-         */
-        Map<String, Object> columnDef = (Map<String, Object>)e.getValue();
-        for (var def: columnDef.entrySet())
-          builder.metadata(def.getKey(), def.getValue().toString());
-      } else {
-        /*
-         * Columns.
-         */
-        boolean notNull = false;
-        String expression = null;
-        Map<String, String> attrs = new LinkedHashMap<>();
-        String type = null;
+  protected CreateTableBuilder builder(Database db) {
+    return new CreateTableBuilder(new Context(db.structure()));
+  }
 
-        if (e.getValue() instanceof String ex) {
-          expression = ex;
-        } else {
-          Map<String, Object> columnDef = (Map<String, Object>)e.getValue();
-          for (var def: columnDef.entrySet()) {
-            switch(def.getKey()) {
-              case "type", "_type" -> type    = (String)def.getValue();
-              case "required"      -> notNull = (Boolean)def.getValue();
-              case "expression"    -> expression = def.getValue().toString();
-              default              -> attrs.put(def.getKey(), def.getValue().toString());
-            }
-          }
-        }
-        if (attrs.containsKey(UNIQUE)) {
-          builder.unique(columnName);
-        }
-        if (attrs.containsKey(LINK_TABLE)
-         && attrs.containsKey(LINK_CODE)) {
-          builder.foreignKey(columnName,
-                             removeQuotes(attrs.get(LINK_TABLE)),
-                             removeQuotes(attrs.get(LINK_CODE)));
-        }
-
-        Attr[] attributes = attrs.entrySet().stream()
-                                 .map(a -> new Attr(a.getKey(), a.getValue()))
-                                 .toArray(Attr[]::new);
-        if (type == null && expression != null) {
-          builder.derivedColumn(columnName, expression, attributes);
-        } else if (type == null) {
-          throw new IllegalArgumentException("Type not specified for non-derived " + columnName);
-        } else {
-          builder.column(columnName, type, notNull, expression, attributes);
-        }
-      }
+  @Override
+  protected void processAttributes(CreateTableBuilder  builder,
+                                   String              columnName,
+                                   Map<String, String> attrs) {
+    if (attrs.containsKey(UNIQUE)) {
+      builder.unique(columnName);
     }
-    try (EsqlConnection con = db.esql()) {
-      con.exec(builder.build());
+    if (attrs.containsKey(LINK_TABLE)
+      && attrs.containsKey(LINK_CODE)) {
+      builder.foreignKey(columnName,
+                         removeQuotes(attrs.get(LINK_TABLE)),
+                         removeQuotes(attrs.get(LINK_CODE)));
     }
-    return get(db, name);
   }
 
   @Override