+++ src/main/java/ma/vi/esql/syntax/define/Create.java
@@ -1,0 +1,9 @@
+package ma.vi.esql.syntax.define;
+
+/**
+ * Tagging interface for create statements.
+ *
+ * @author Vikash Madhow (vikash.madhow@gmail.com)
+ */
+public interface Create {
+}